requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
Flask>=2.0.0
